Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/eventsource/eventsource-constructor-non-same-origin-expected.txt

CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: EventSource cannot load ftp://example.not/ due to access control checks.
CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: EventSource cannot load about:blank due to access control checks.
CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: EventSource cannot load mailto:whatwg@awesome.example due to access control checks.
CONSOLE MESSAGE: Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: EventSource cannot load javascript:alert('FAIL') due to access control checks.

PASS EventSource: constructor (act as if there is a network error) (ftp://example.not/)
PASS EventSource: constructor (act as if there is a network error) (about:blank)
PASS EventSource: constructor (act as if there is a network error) (mailto:whatwg@awesome.example)
PASS EventSource: constructor (act as if there is a network error) (javascript:alert('FAIL'))
